Dreamboxes: Smarter compression settings, enable dm8000 WiFi and DVD
Don't compress the "boot" partition, it only contains compressed data
like the .gz kernel and image files.
The "big" boxes have plenty flash, so using LZO is the smarter choice.
The image will be 10% larger, but loads considerably faster (not as
fast as UBIFS would, which is also a nice avenue to explore).

The dm8000 wifi support is now built-in to the kernel since you cannot
buy the dm8000 without it. Also CD/DVD filesystems are supported in
the kernel now.

(note: Only tested on dm8000, if i did something wrong here, other DM
boxes might not even boot up tomorrow)