tuxtxt-enigma2: Fix build and renovate recipe
CONFFILES doesn't support wildcards, and the files need to be in the package.

There's no reason to override the "dbg" and "dev" contents, so do not mention
them (by default, they'll pick up the .debug stuff). This also fixes a build
warning that the .so and .la files weren't being deployed in any package.