mdev-mount.sh: Tell enigma about non-mountable devices
If a disk doesn't contain a filesystem, it's still a good thing to
tell the UI about it. If you attach an unpartitioned or otherwise
unrecognized disk, you still want to be able to format it. (*)

This may fix some reports about "missing" USB drives, since they
wouldn't show up in Enigma2 if they weren't yet there at startup.

(*) Note: E2 could just listen on a netlink socket for this.