distro: This is OpenPLi 4
On purpose I did not create a "4.0" or "3.1". I call this version "4" because
it's newer than 3, and I don't want to argue about whether this is a minor or
major step.

Also, dropping the "suffix" version might make it clear that this version
number bears no relation whatsoever with other OE environments.