settings-autorestore: Move startup from S20 to S31
As the startup speed increases, the local disks will announce later,
so push it ahead to right before the network initialization.
This already fails on the vusolo2 where a USB stick mounts only after
reaching runlevel 3. So it seems the restore should move into E2,
right before starting the wizard.