linux (>= 3.8): revert recent change in r4k_dma_cache_inv
This change in the cache instruction handling is causing
problems with various dvb-usb drivers.
According to the original commit log, this might be caused
by the use of badly aligned dma addresses, in those drivers.
However, since we did not have any noticable problems with
the old handling, we simply revert for now.