dm800se/defconfig: Run through menuconfig and remove UBI and fb console
Run menuconfig to make the defconfig file reflect what is actually being
done.

Remove UBI, since it isn't being used and makes the kernel too big to
boot.

Removed the framebuffer console, it isn't being used and adds several
fonts to the kernel, making it bigger.

Add a check into the build so that a bigger kernel will trigger a build
failure.