Use busybox ftpd instead of vsftpd (don't ever forward port 21)
vsftpd has troubles with blank passwords. Our version was outdated
so should have been removed anyway.

Busybox provides a nice package that allows you to get and put files
on the system with little overhead, so let's try that. Smaller is
sometimes better.

This applet does not have any kind of security whatsoever.

It would be nice to be able to install vsftpd as an option for users
who want more. I promise I'll look into that later.