Add vsftpd configuration for OpenPLi boxes
Add a low-security configuration for vsftpd that works. A bit of hacking though
to remove the initscripts and useradd stuff from the package.

Re-activating the "vsftpd" line in /etc/inetd.conf will use the vsftpd server
instead of the busybox one.

This version also bails out if your password is empty.

This only adds the recipe, but doesn't build nor activate this server.