nfs-utils: Fix and simplify broken startup script
This line was causing the script to abort:
grep -q nfsd /proc/filesystems
Apparently we don't have it there, but the server runs fine.

The script was overly complex, so simplify it (no need to check
for the existence of a file, just use it and it'll fail without
any help from shell scripts) to reduce startup time.