opkg: hack to use http if_modified_since
This is a crazy hack to avoid repeatatly downloading big Packages.gz
files by opkg.
Remove gzip handling from opkg and perform gunzip from an external
script. To deal with expectations of opkg some weird moving and
copying around is performed by the script to make opkg happy.
All this should actually be part of the core opkg but is to
complicated to do in a few minutes.