Update to OE-core master
This causes a LOT of package rebuilds. Homebrew images are advised to remove
the "tmp" directory and let it start compiling all over.
As this will touch hundreds of packages, some boxes may run into trouble
when flash space runs low. There's no cure for that except re-flashing, so
a quick autobackup of your settings before upgrading is - as always - advised.

Add a bunch of "brokensep" directives to packages that won't build outside
their source dirs.

Fix libxmlccwrap to use pkgconfig so that it finds the includes it needs.

Remove SWIG recipe, oe-core already provides a newer version.