enigma2-plugins: Don't store anything in sysroot
If the recipe builds libraries and exports include files, they're not
intended for further distribution. Force sysroot_stage_all to be a
no-op, so that nothing ends up in the sysroot where it may conflict
with other recipes.

This prevents build failures when multiple plugin recipes provide
the same files.