openpli-enigma2-feed: trigger package index creation
The openpli-enigma2-feed recipe doesn't build a rootfs, so it also does
not trigger the package indexer. Trigger this by simply "inheriting"
from the package-index recipe. It also makes the recipe quicker by
removing lots of steps that aren't needed.

This should resolve the issue as reported by 'ims' on the forum:
http://forums.openpli.org/topic/35374-master-next-to-become-master/page-2#entry455323