dreambox.inc: Don't override native linux-libc-headers version
I see no reason why only the dreambox machines would need the native
binaries to be compiled with old headers, none of the other boxes do
that. Remove the directive.

Just a small step on the way to completely removing the
linux-libc-headers overrides, these are causing the boxes to have
different package versions.