OE-core upgrade for gitpkgv fix
This upgrades everything, so expect a 300+ packages upgrade and
having to run it three times if you do an online upgrade. The box
will survive, but just reflashing is going to be quicker.

Update openembedded and core components to today's state. This
includes the fix for PR-server breaking gitpkgv.

Apparently a newer git version, rename the bbappend to match.

The kernel build recipes have undergone massive changes in newer
OpenEmbedded versions. This results in build failures like this:
| make: *** No rule to make target `oldconfig'.  Stop.
To resolve the issue, move the assignment of "S" to after the
"inherit kernel" statement, and remove the manual defconfig
handling, OE already takes care of that.