Remove closed-source vendor specific "gst-plugin-libxt"
"gst-plugin-libxt" has an incompatible license, and only contains a binary object
that is apparently box specific. At best, this could be a vendor specific plugin.
It should not be in a GPLv2 licensed open source project, though everyone is free
to do with it as he pleases in his own home.