Remove openpli-unstable-versions.inc and reduce AUTOREV usage
openpli-unstable-versions.inc was intended to help "pin" a release, but
we never did this, nor does it actually help in managing versions.

A few packages in the list did not exist any longer.

Move the SRCREV values into the packages. This also makes it easier to
override them in a local.conf or other constructs than editing the file.
It also allows the recipes to be used in an other distro, since depending
on a specific distro just to get recipes to work is not a good thing.

AUTOREV causes the bitbake parser to contact the remote server to get
the version number. This makes the parse stage take much longer and also
makes the complete build fail if the server is offline. Some servers
appear to have filters, and will fail if you contact the same repo
several times in a short timespan.
It also makes it virtually impossible to reproduce a past build, and
causes several packages to "lie" about their version.

Reduce the usage of AUTOREV by pinning them to the current HEAD. This does
indeed mean that you'll have to make a one-line change if you want a newer
version of that repository.

This will help in getting more reliable and faster builds.