Revert "vuzero/vusolose: add pau."
This reverts commit e810bc3f4b4dabc00fe98c47c96e32dd840753f9.

Vendor specific stuff should go in the bsp layer,
several mechanisms exist to add a vendor-specific
package to the image or feed.

In this case, the package 'pau' could be added to
MACHINE_ESSENTIAL_EXTRA_RDEPENDS or MACHINE_ESSENTIAL_EXTRA_RRCOMMENDS,
in order to have it installed in the image.