enigma2 plugins: do not enforce "enigma2" dependency
Plugins should not RDEPEND on the package they are plugging into.
Two main reasons:
- If someone creates an "enigma3" that can run these plugins, it will
be next to impossible to install such plugins.
- If enigma2 RRECOMMENDs a plugin, this creates a cyclic dependency.

Only RDEPEND on enigma2 if a plugin actually links to internal E2 code,
which is something that OpenEmbedded will detect automatically.

Using extra_depends='' in do_split_packages removes the default dependency
on the main PN.