fakelocale pl_PL lc_time: change month names to genitive, for grammatically correct displaying of date
in Polish, as in other slavic languages, a full month name in date presentation should be in genitive, using nominative instead is an obvious grammatical error, which now appears on almost every GUI screen.
Unfortunately glibc locale do not allow alternative forms for full month names (as opposed to e.g. FreeBSD, where the genitive is even the basic form), therefore need to choose only one form.
The stand alone month names, which alone require the use of nominative (and uppercasing btw), are definitely of much much less frequent use in Polish, in general.
Moreover in ClockToText converter %B is only used together with a day number, and, actually, I never saw a stand alone month name used anywhere in the image or plugin, hence this workaround to fix a common issue could be considered reasonable and harmless.

Signed-off-by: Erik Slagter <erik@openpli.org>