pd1loi skin: remove for now.
It breaks the build because it refers to an invalid license.

But moreover, it's completely wrong. It will have to be re-orded
so it acts like a normal skin/plugin. No install scripts, no
tar files, no reboots... Just a mirror of the files to be installed
and then put that in the package.