busybox mdev-mount.sh: mount first internal device on /media/hdd only if label not set
This allows using a label for mounting an internal hdd, but it can break things
in case the internal hdd has a label other than 'hdd'. If the box formatted the
disk, like it should, it will not have a label.

Use tune2fs to change the label to "hdd".

Signed-off-by: Mike Looijmans <milo@openpli.org>
[ML] Changed commit text.