busybox/mdev-mount.sh: Special-case SD/MMC devices. Avoid remounting root.
MMC and SD devices use names like "mmcblk0p1". Handle this specially, to
properly detect the device and call it "SD-card" when removable or "mmc"
when internal.

On the vusolo4k, root is mounted on /dev/root which is a symlink to the
actual root device. Instead of handling symlinks in the mount tab, just
special case this and don't mount the rootfs twice.