Add capability to abort restore settings and auto install
When you install an image by a fresh reflash the auto restore and auto
install was always executed. So when you had an automatic backup on the
root of your harddisk this will always install. Sometimes when the backup
was corrupted you need to disconect the harddisk drive of your box. Now
this is not required anymore. When you put on a root of any connected
media (the harddisk itself, but also the USB drive you use to flash the
image) a file named 'nobackup' the auto restore and auto install will be
aborted so you'll get a clean fresh flash without removing the harddisk
drive.