libjpeg-turbo: workaround for compilation on mips32el-nf.
This is very crude. I can't find a "proper" way to do it in OE.
Volunteers are very wellcome.

Problem: libjpeg uses SIMD instructions, also on mips. The SIMD
instructions are not available on non-FPU SoC's (mips32el-nf) and
they're also not emulated (at least that is gcc's opinion, apparently).

OE calls gcc with the proper flags (--soft-float), so it recognises
it can't use SIMD instructions. The libjpeg lib contains assembly files
that include SIMD instructions, though. This will make gcc bail out
on non-FPU SoC's.

To prevent this, libjpeg-turbo has a config flag "--no-simd" which is
effective, but it isn't set by OE on mips32el-nf. It seems there is no
way to detect (and act on) the fact that we're compiling for non-FPU
from within OE.

In process I found out that the configure script actually does actively
check for SIMD support on the processor, but uses the wrong test for it,
so I doesn't fail when no SIMD is available. Extended the test with a
proper test for actual SIMD acceptance and that fixes the issue.