Remove gstreamer recipes
Remove the gstreamer recipes that were masking OE-core ones. The assumption
is that all 1.6 issues have been fixed in 1.8. If not, we'll bring back the
neccesary patches with bbappends (and hopefully, an upstream status). This
way I can at least focus on getting things to compile instead of patching the
same merge problems over and over again.