Ensure smbd is working when you restore Openpli4 on an Openpli5 image
OpenPLi5 uses a new version of smbd. The smb.conf file from OpenPLi4 is
therefor not compatible anymore with OpenPLi5. This images backups the
smb.conf from OpenPLi5, thich is recognized by a "netbios name = %h"
string.
After the restore it is checked if we have a backed up smb.conf file,
which confirms we are on OpenPLi5, then it checks if the restored smb.conf
was restored from OpenPLi4. If so it restored the backup made from the
Openpli5 image.