busybox/mdev-mount.sh: Workaround to mount ext2/3 as ext4 with util-linux mount
The mount command from util-linux tries to mount ext2 partitions as "ext2" but
that filesystem no longer exists in the kernel. Apparently the util-linux and kernel
folks are still fighting of who's responsible for this sillyness, so it still
fails to mount.

Example:
/bin/mount.util-linux -v -t auto /dev/sda1 /media/usb
mount.util-linux: unknown filesystem type 'ext2'
/bin/mount.util-linux -v -t auto,ext4 /dev/sda1 /media/usb
mount.util-linux: /dev/sda1 mounted on /media/usb.

The second form instructs mount to try "ext4" explicitly (which will mount ext2/3
as well) when auto fails. Busybox mount doesn't seem to mind the "auto,ext4" appended
so this is a working workaround for the bug in util-linux mount. Busybox had no such
trouble because its "auto" strategy is to simple try all possible filesystems.