empty rootfs before do_rootfs() starts, not when it finishes
this leaves tmp/rootfs present for subsequent tasks like do_image()
which may need files from the rootfs, but current isn't able to access
those because rm_work immediately cleans the directory after do_rootfs()
is finished. This is particularly a problem when bitbake decides not to
call do_rootfs(), leaving the rootfs directory empty, and causing the
do_image() tasks to fail.