busybox: Enable ASH_CMDCMD for built-in 'command'
See openembedded-core commit e41c90b852167bbcb434da57b84280e855acae33

It is common for *ash shells to have 'command' available as a built-in
function. POSIX 2008 also documents the availability of this command.

Additionally the /etc/profile of base-files requires this command to be
available as of commit e77cdb7611 ("base-files: profile: Do not assume
that the tty command exists"). If it is not available the following
message is output during login on a image using busybox.

  -sh: command: not found