initscripts: exit when script is called with 'stop' argument
These simple initscripts do not check their arguments, they expect to be
executed with the 'start' argument.
However, during opkg upgrade for instance, the update-rc.d bbclass
attempts to stop the service in preinst hook.

This might have undesired sideeffects for the upgrade process
(needlessly delaying it, or possibly even aborting it)