Revert "rocko: switch from busybox-mdev to (e)udev."
This reverts commit 95c0a9e16ca11b7dfa684719fe7c4609d0184f4f.

There's a bit more to replacing mdev than just installing eudev. It breaks automounting
disks (which could be partially restored with "udev-extraconf"), system shutdown,
formatting and checking disks from enigma, and probably more. Let's focus on a working
setup first, mdev may not be perfect but it is not broken.