opkg: add patch for more symlink workarounds.
The previous version of the patch cared for the case where a symlink to
a directory is replacing a directory and v.v. but there appeared to be
another case, where a symlink to directory is replacing a symlink to a
directory, which was not caught by the workaround.

Now if existing file and new file both are a symlink, allow it anyway,
and skip any other replacement check, it is always allowed to replace a
symlink with another symlink.