kodi_18: do not unconditionally skip dp_package_qa
from the build tests of all models only the zgemma models need to
skip this check, for the moment.
This happens when the solibs lack SONAME or are not versioned (or in
case of PRIVATE_LIBS).
Need more checks for this BSP.
One specific runtime dependency might be added to silence the QA.

Signed-off-by: Andrea Adami <andrea.adami@gmail.com>