feed: Remove plugins which doesn't compile
The main goal for now is Enigma2 and e2plugins.
Remove these plugins to be able to build the feed at least so
that users have a softcam, settings list and some plugin.

Most of them requires a little change it seems, like missing parenthesises
with print function ins setup.py file.