python3: add dummy pycache files
Building a patched version of python3 with byte-code files saved to their legacy locations and names causes the following errors:

| rm: cannot remove '/.../build/tmp/work/x86_64-linux/python3-native/3.9.9-r0/recipe-sysroot-native/usr/lib/python3.9/test/__pycache__/test_range.cpython*': No such file or directory
| rm: cannot remove '/.../build/tmp/work/x86_64-linux/meson-native/0.57.1-r1/recipe-sysroot-native/usr/lib/python*/site-packages/mesonbuild/dependencies/__pycache__/mpi.cpython*': No such file or directory

That happens because OE expects those file in __pycache__ folder and now due to lecaly byte-code are missing.
So in order to make rm happy, add a do_install_prepend to create touch the files for rm.